+++ tests/testthat/test-check_synapse_login.R
@@ -1,4 +1,0 @@
-test_that("Invalid function parameter", {
-  expect_error(.is_connected_to_genie("GENIE"))
-})
-